custom_actions/requirements.txt
pyyaml
requests
pyfiglet